extracted oauth creation flow from OAuth2Manager and moved logic into cms-cli
